refactor(schemas): extract validate_typeddict to src/schemas/validator.py
Move the hand-rolled TypedDict validator out of tests/contract/ so it can
be reused by the upcoming dev-mode response-schema middleware (JTN-664).
Zero behavior change; contract tests still drive through the same
implementation.
